Moved Generator Menu out of GeneratorWindow.
