+++ UnityProject/Assets/Editor/LevelGenerator/GUI/GeneratorMenu.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0ba58f56c962543d8aa9d05a96c7fb5d
+timeCreated: 1503415541
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
