+++ UnityProject/Assets/Editor/LevelGenerator/GUI/Menu.cs
@@ -2,3 +2,14 @@
 {
 	void Menu();
 }
+
+public class GenData
+{
+	public string name = "name";
+	public string x = "0";
+	public string y = "0";
+	public int idx = 0;
+	public string width = "";
+	public string height = "";
+}
+
