+++ UnityProject/Assets/GeneratorMenu.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace AssemblyCSharpEditor
+{
+	public class GeneratorMenu
+	{
+		public GeneratorMenu ()
+		{
+		}
+	}
+}
+
